requirements.txt
joblib==1.4.2
streamlit==1.38.0
numpy==1.26.4
pillow==10.4.0
matplotlib==3.9.2
#matplotlib-inline==0.1.7
seaborn==0.13.2
#tflite-runtime
#ailia_tflite
tf-nightly-cpu
#tensorflow-cpu==1.14
#keras~=3.0
#tf-nightly